// src/app/_restTarjetaCredito/model/tarjetaCredito.ts
/**
 * Servicio TarjetasCredito Kallsonys
 * Servicio para la gestión de tarjetas de credito en OMS Kallsonys
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Tarjetas de credito de la plataforma Kallsonys
 */
export interface TarjetaCredito { 
    idtarjeta?: number;
    numtarjeta?: string;
    franquicia?: string;
    cantidad?: number;
    fechacreacion?: string;
    estado?: string;
    idcliente?: number;
}